Update the rest of the trade states
