+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/trade_state/states/BuyerState4.java
@@ -18,27 +18,19 @@
 package bisq.desktop.main.content.bisq_easy.open_trades.trade_state.states;
 
 import bisq.bonded_roles.explorer.ExplorerService;
-import bisq.bonded_roles.explorer.dto.Output;
 import bisq.chat.bisqeasy.open_trades.BisqEasyOpenTradeChannel;
-import bisq.common.monetary.Coin;
+import bisq.chat.bisqeasy.open_trades.BisqEasyOpenTradeChannelService;
+import bisq.chat.bisqeasy.open_trades.BisqEasyOpenTradeSelectionService;
 import bisq.desktop.ServiceProvider;
 import bisq.desktop.common.Browser;
-import bisq.desktop.common.threading.UIScheduler;
-import bisq.desktop.common.threading.UIThread;
+import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.controls.MaterialTextField;
-import bisq.desktop.main.content.bisq_easy.components.WaitingAnimation;
-import bisq.desktop.main.content.bisq_easy.components.WaitingState;
 import bisq.desktop.components.controls.WrappingText;
 import bisq.desktop.components.overlay.Popup;
+import bisq.desktop.main.content.bisq_easy.open_trades.trade_state.OpenTradesUtils;
 import bisq.i18n.Res;
-import bisq.presentation.formatters.AmountFormatter;
-import bisq.trade.TradeException;
 import bisq.trade.bisq_easy.BisqEasyTrade;
 import de.jensd.fx.fontawesome.AwesomeIcon;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
 import javafx.geometry.Insets;
 import javafx.scene.control.Button;
 import javafx.scene.layout.HBox;
@@ -46,8 +38,6 @@ import javafx.scene.layout.VBox;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-
-import java.util.concurrent.TimeUnit;
 
 @Slf4j
 public class BuyerState4 extends BaseState {
@@ -63,12 +53,15 @@ public class BuyerState4 extends BaseState {
 
     private static class Controller extends BaseState.Controller<Model, View> {
         private final ExplorerService explorerService;
-        private UIScheduler scheduler;
+        private final BisqEasyOpenTradeChannelService bisqEasyOpenTradeChannelService;
+        private final BisqEasyOpenTradeSelectionService bisqEasyOpenTradeSelectionService;
 
         private Controller(ServiceProvider serviceProvider, BisqEasyTrade bisqEasyTrade, BisqEasyOpenTradeChannel channel) {
             super(serviceProvider, bisqEasyTrade, channel);
 
             explorerService = serviceProvider.getBondedRolesService().getExplorerService();
+            bisqEasyOpenTradeChannelService = serviceProvider.getChatService().getBisqEasyOpenTradeChannelService();
+            bisqEasyOpenTradeSelectionService = serviceProvider.getChatService().getBisqEasyOpenTradesSelectionService();
         }
 
         @Override
@@ -84,88 +77,41 @@ public class BuyerState4 extends BaseState {
         @Override
         public void onActivate() {
             super.onActivate();
-
             model.setTxId(model.getBisqEasyTrade().getTxId().get());
-            model.setBtcAddress(model.getBisqEasyTrade().getBtcAddress().get());
-            model.getBtcBalance().set("");
-            model.getConfirmationState().set(Res.get("bisqEasy.tradeState.info.phase4.balance.help.explorerLookup"));
-            requestTx();
         }
 
         @Override
         public void onDeactivate() {
             super.onDeactivate();
-            if (scheduler != null) {
-                scheduler.stop();
-                scheduler = null;
-            }
         }
 
-        public void openExplorer() {
+        private void onLeaveChannel() {
+            new Popup().feedback(Res.get("bisqEasy.openTrades.closeTrade.warning.completed"))
+                    .actionButtonText(Res.get("confirmation.yes"))
+                    .onAction(() -> {
+                        channelService.leaveChannel(model.getChannel());
+                        bisqEasyTradeService.removeTrade(model.getBisqEasyTrade());
+                        selectionService.getSelectedChannel().set(null);
+                    })
+                    .closeButtonText(Res.get("confirmation.no"))
+                    .show();
+        }
+
+        private void onExportTrade() {
+            OpenTradesUtils.exportTrade(model.getBisqEasyTrade(), getView().getRoot().getScene());
+        }
+
+        private void openExplorer() {
             ExplorerService.Provider provider = explorerService.getSelectedProvider().get();
             String url = provider.getBaseUrl() + provider.getTxPath() + model.getTxId();
             Browser.open(url);
-        }
-
-        private void onComplete() {
-            try {
-                bisqEasyTradeService.btcConfirmed(model.getBisqEasyTrade());
-            } catch (TradeException e) {
-                new Popup().error(e).show();
-            }
-        }
-
-        private void requestTx() {
-            explorerService.requestTx(model.getTxId())
-                    .whenComplete((tx, throwable) -> {
-                        UIThread.run(() -> {
-                            if (scheduler != null) {
-                                scheduler.stop();
-                            }
-                            if (throwable == null) {
-                                model.btcBalance.set(
-                                        tx.getOutputs().stream()
-                                                .filter(output -> output.getAddress().equals(model.getBtcAddress()))
-                                                .map(Output::getValue)
-                                                .map(Coin::asBtcFromValue)
-                                                .map(e -> AmountFormatter.formatAmountWithCode(e, false))
-                                                .findAny()
-                                                .orElse(""));
-                                model.getIsConfirmed().set(tx.getStatus().isConfirmed());
-                                if (tx.getStatus().isConfirmed()) {
-                                    onConfirmed();
-                                } else {
-                                    model.getConfirmationState().set(Res.get("bisqEasy.tradeState.info.phase4.balance.help.notConfirmed"));
-                                    scheduler = UIScheduler.run(this::requestTx).after(20, TimeUnit.SECONDS);
-                                }
-                            } else {
-                                model.getConfirmationState().set(Res.get("bisqEasy.tradeState.info.phase4.txId.failed"));
-                                log.warn("Transaction lookup failed", throwable);
-                            }
-                        });
-                    });
-        }
-
-        private void onConfirmed() {
-            model.getConfirmationState().set(Res.get("bisqEasy.tradeState.info.phase4.balance.help.confirmed"));
-            sendSystemMessage(Res.get("bisqEasy.tradeState.info.phase4.systemMessage", model.getFormattedBaseAmount(), model.btcAddress));
-            try {
-                bisqEasyTradeService.btcConfirmed(model.getBisqEasyTrade());
-            } catch (TradeException e) {
-                new Popup().error(e).show();
-            }
         }
     }
 
     @Getter
     private static class Model extends BaseState.Model {
         @Setter
-        protected String btcAddress;
-        @Setter
         protected String txId;
-        private final StringProperty btcBalance = new SimpleStringProperty();
-        private final StringProperty confirmationState = new SimpleStringProperty();
-        private final BooleanProperty isConfirmed = new SimpleBooleanProperty();
 
         protected Model(BisqEasyTrade bisqEasyTrade, BisqEasyOpenTradeChannel channel) {
             super(bisqEasyTrade, channel);
@@ -173,29 +119,35 @@ public class BuyerState4 extends BaseState {
     }
 
     public static class View extends BaseState.View<Model, Controller> {
-        private final Button button;
-        private final MaterialTextField txId, btcBalance;
-        private final WaitingAnimation waitingAnimation;
+        private final Button leaveButton, exportButton;
+        private final MaterialTextField quoteAmount, baseAmount;
+        private final MaterialTextField txId;
 
         private View(Model model, Controller controller) {
             super(model, controller);
 
+            WrappingText headline = FormUtils.getHeadline(Res.get("bisqEasy.tradeState.info.buyer.phase4.headline"));
+
+            exportButton = new Button(Res.get("bisqEasy.tradeState.info.phase5.exportTrade"));
+            leaveButton = new Button(Res.get("bisqEasy.tradeState.info.phase5.leaveChannel"));
+            leaveButton.getStyleClass().add("outlined-button");
+            quoteAmount = FormUtils.getTextField(Res.get("bisqEasy.tradeState.info.buyer.phase4.quoteAmount"), "", false);
+            baseAmount = FormUtils.getTextField(Res.get("bisqEasy.tradeState.info.buyer.phase4.baseAmount"), "", false);
+
             txId = FormUtils.getTextField(Res.get("bisqEasy.tradeState.info.phase4.txId"), "", false);
             txId.setIcon(AwesomeIcon.EXTERNAL_LINK);
             txId.setIconTooltip(Res.get("bisqEasy.tradeState.info.phase4.txId.tooltip"));
-            btcBalance = FormUtils.getTextField(Res.get("bisqEasy.tradeState.info.buyer.phase4.balance"), "", false);
-            btcBalance.setHelpText(Res.get("bisqEasy.tradeState.info.phase4.balance.help.explorerLookup"));
-            btcBalance.setPromptText(Res.get("bisqEasy.tradeState.info.buyer.phase4.balance.prompt"));
 
-            button = new Button(Res.get("bisqEasy.tradeState.info.phase4.buttonText"));
-            VBox.setMargin(button, new Insets(5, 0, 5, 0));
+            HBox buttons = new HBox(leaveButton, Spacer.fillHBox(), exportButton);
 
-            waitingAnimation = new WaitingAnimation(WaitingState.BITCOIN_CONFIRMATION);
-            WrappingText headline = FormUtils.getHeadline(Res.get("bisqEasy.tradeState.info.buyer.phase4.headline"));
-            WrappingText info = FormUtils.getInfo(Res.get("bisqEasy.tradeState.info.buyer.phase4.info"));
-            HBox waitingInfo = createWaitingInfo(waitingAnimation, headline, info);
-
-            root.getChildren().addAll(waitingInfo, txId, btcBalance, button);
+            VBox.setMargin(headline, new Insets(0, 0, 5, 0));
+            VBox.setMargin(buttons, new Insets(5, 0, 5, 0));
+            root.getChildren().addAll(
+                    headline,
+                    quoteAmount,
+                    baseAmount,
+                    txId,
+                    buttons);
         }
 
         @Override
@@ -204,28 +156,21 @@ public class BuyerState4 extends BaseState {
 
             txId.setText(model.getTxId());
 
-            button.defaultButtonProperty().bind(model.isConfirmed);
-            btcBalance.textProperty().bind(model.getBtcBalance());
-            btcBalance.helpHelpProperty().bind(model.getConfirmationState());
+            quoteAmount.setText(model.getFormattedQuoteAmount());
+            baseAmount.setText(model.getFormattedBaseAmount());
 
-            button.setOnAction(e -> controller.onComplete());
+            leaveButton.setOnAction(e -> controller.onLeaveChannel());
+            exportButton.setOnAction(e -> controller.onExportTrade());
             txId.getIconButton().setOnAction(e -> controller.openExplorer());
-
-            waitingAnimation.play();
         }
 
         @Override
         protected void onViewDetached() {
             super.onViewDetached();
 
-            button.defaultButtonProperty().unbind();
-            btcBalance.textProperty().unbind();
-            btcBalance.helpHelpProperty().unbind();
-
-            button.setOnAction(null);
+            leaveButton.setOnAction(null);
+            exportButton.setOnAction(null);
             txId.getIconButton().setOnAction(null);
-
-            waitingAnimation.stop();
         }
     }
 }
