+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/trade_state/states/SellerState1.java
@@ -112,6 +112,7 @@ public class SellerState1 extends BaseState {
         @Override
         public void onDeactivate() {
             super.onDeactivate();
+
             accountsPin.unbind();
             selectedAccountPin.unbind();
             model.getButtonDisabled().unbind();
@@ -123,8 +124,8 @@ public class SellerState1 extends BaseState {
                 new Popup().warning(Res.get("validation.tooLong", UserProfile.MAX_LENGTH_STATEMENT)).show();
                 return;
             }
-            String message = Res.get("bisqEasy.tradeState.info.seller.phase1.systemMessage", paymentAccountData);
-            sendSystemMessage(message);
+            //String message = Res.get("bisqEasy.tradeState.info.seller.phase1.systemMessage", paymentAccountData);
+            //sendSystemMessage(message);
             try {
                 bisqEasyTradeService.sellerSendsPaymentAccount(model.getBisqEasyTrade(), paymentAccountData);
             } catch (TradeException e) {
