+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/BisqEasySellerAsTakerProtocol.java
@@ -43,8 +43,34 @@ public class BisqEasySellerAsTakerProtocol extends BisqEasyProtocol {
                 .run(BisqEasyTakeOfferResponseHandler.class)
                 .to(TAKER_RECEIVED_TAKE_OFFER_RESPONSE);
 
+        // Option 1: Seller sends first account data, then buyer sends btc address
         addTransition()
                 .from(TAKER_RECEIVED_TAKE_OFFER_RESPONSE)
+                .on(BisqEasyAccountDataEvent.class)
+                .run(BisqEasyAccountDataEventHandler.class)
+                .to(SELLER_SENT_ACCOUNT_DATA);
+
+        addTransition()
+                .from(SELLER_SENT_ACCOUNT_DATA)
+                .on(BisqEasyBtcAddressMessage.class)
+                .run(BisqEasyBtcAddressMessageHandler.class)
+                .to(SELLER_RECEIVED_BTC_ADDRESS);
+
+        addTransition()
+                .from(SELLER_RECEIVED_BTC_ADDRESS)
+                .on(BisqEasyConfirmFiatSentMessage.class)
+                .run(BisqEasyConfirmFiatSentMessageHandler.class)
+                .to(SELLER_RECEIVED_FIAT_SENT_CONFIRMATION);
+
+        // Option 2: Buyer sends first btc address, then seller sends account data
+        addTransition()
+                .from(TAKER_RECEIVED_TAKE_OFFER_RESPONSE)
+                .on(BisqEasyBtcAddressMessage.class)
+                .run(BisqEasyBtcAddressMessageHandler.class)
+                .to(SELLER_RECEIVED_BTC_ADDRESS);
+
+        addTransition()
+                .from(SELLER_RECEIVED_BTC_ADDRESS)
                 .on(BisqEasyAccountDataEvent.class)
                 .run(BisqEasyAccountDataEventHandler.class)
                 .to(SELLER_SENT_ACCOUNT_DATA);
@@ -55,14 +81,9 @@ public class BisqEasySellerAsTakerProtocol extends BisqEasyProtocol {
                 .run(BisqEasyConfirmFiatSentMessageHandler.class)
                 .to(SELLER_RECEIVED_FIAT_SENT_CONFIRMATION);
 
+        // from here it's the same
         addTransition()
                 .from(SELLER_RECEIVED_FIAT_SENT_CONFIRMATION)
-//                .on(BisqEasyBtcAddressMessage.class)
-//                .run(BisqEasyBtcAddressMessageHandler.class)
-//                .to(SELLER_RECEIVED_BTC_ADDRESS);
-//
-//        addTransition()
-//                .from(SELLER_RECEIVED_BTC_ADDRESS)
                 .on(BisqEasyConfirmFiatReceiptEvent.class)
                 .run(BisqEasyConfirmFiatReceivedEventHandler.class)
                 .to(SELLER_CONFIRMED_FIAT_RECEIPT);
@@ -116,11 +137,11 @@ public class BisqEasySellerAsTakerProtocol extends BisqEasyProtocol {
                 .on(BisqEasyCancelTradeMessage.class)
                 .run(BisqEasyCancelTradeMessageHandler.class)
                 .to(CANCELLED);
-//        addTransition()
-//                .from(SELLER_RECEIVED_BTC_ADDRESS)
-//                .on(BisqEasyCancelTradeMessage.class)
-//                .run(BisqEasyCancelTradeMessageHandler.class)
-//                .to(CANCELLED);
+        addTransition()
+                .from(SELLER_RECEIVED_BTC_ADDRESS)
+                .on(BisqEasyCancelTradeMessage.class)
+                .run(BisqEasyCancelTradeMessageHandler.class)
+                .to(CANCELLED);
         addTransition()
                 .from(SELLER_CONFIRMED_FIAT_RECEIPT)
                 .on(BisqEasyCancelTradeMessage.class)
@@ -143,11 +164,11 @@ public class BisqEasySellerAsTakerProtocol extends BisqEasyProtocol {
                 .on(BisqEasyCancelTradeEvent.class)
                 .run(BisqEasyCancelTradeEventHandler.class)
                 .to(CANCELLED);
-//        addTransition()
-//                .from(SELLER_RECEIVED_BTC_ADDRESS)
-//                .on(BisqEasyCancelTradeEvent.class)
-//                .run(BisqEasyCancelTradeEventHandler.class)
-//                .to(CANCELLED);
+        addTransition()
+                .from(SELLER_RECEIVED_BTC_ADDRESS)
+                .on(BisqEasyCancelTradeEvent.class)
+                .run(BisqEasyCancelTradeEventHandler.class)
+                .to(CANCELLED);
         addTransition()
                 .from(SELLER_CONFIRMED_FIAT_RECEIPT)
                 .on(BisqEasyCancelTradeEvent.class)
