+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/BisqEasyTradeState.java
@@ -37,7 +37,7 @@ public enum BisqEasyTradeState implements State {
     SELLER_RECEIVED_FIAT_SENT_CONFIRMATION,
 
     BUYER_SENT_BTC_ADDRESS,
-    //SELLER_RECEIVED_BTC_ADDRESS,
+    SELLER_RECEIVED_BTC_ADDRESS,
 
     SELLER_CONFIRMED_FIAT_RECEIPT,
     BUYER_RECEIVED_SELLERS_FIAT_RECEIPT_CONFIRMATION,
